Plugged the main layout on iOS (still need to plug the agenda)
