--- iosApp/RobotConf/UI/About/AboutView.swift
@@ -1,14 +1,0 @@
-//
-//  Copyright © 2020 Paris Android User Group. All rights reserved.
-//
-
-import SwiftUI
-import shared
-
-struct AboutView: UIViewControllerRepresentable {
-    func makeUIViewController(context: Context) -> UIViewController {
-        return AboutViewController()
-    }
-
-    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {}
-}
